requirements.txt
# Core AI & LangChain
langchain
langchain-openai
langchain-community
langchain-google-genai

# Google Gemini
google-generativeai

# LLM / API
openai
python-dotenv

# Voice I/O
pyttsx3
vosk
SpeechRecognition
gTTS
playsound==1.2.2
# simpleaudio  # Not used, can remove if not needed
# pydub        # Not used, can remove if not needed
# TTS          # Not used, can remove if not needed
# resemblyzer  # Not used, can remove if not needed

# GUI
pillow
tk

# Web tools
requests
beautifulsoup4
lxml

# System & Automation
psutil
pyautogui
keyboard
mouse

# Time
pytz

# Office Tools
python-docx
python-pptx
XlsxWriter
openpyxl


# Email & Google API
google-api-python-client
google-auth
google-auth-oauthlib
pyaudio

# Networking
whois
speedtest-cli


# File System & Utilities
# shutilwhich  # Not used, can remove if not needed
# pyyaml       # Not used, can remove if not needed


# Vision / OpenCV
opencv-python
numpy
chromadb
pigpio
Adafruit_DHT
adafruit-circuitpython-dht
spidev
# python-lirc  # Requires legacy CPython headers; optional and omitted by default

# Security
cryptography
paramiko


# Misc / Helper
# rich      # Not used, can remove if not needed
# tabulate  # Not used, can remove if not needed

# optional helpers
# urllib3   # Not used, can remove if not needed
# certifi   # Not used, can remove if not needed
